Moves Percent tests to the crate with the type definition (#1714)
